+++ CHANGELOG.md
@@ -71,12 +71,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - validateSignature method to webhooks  
 
-## [2.1.0] - 2020-06-04
+## [2.1.0] - 2020-06-05
   
 ### Added
   
   - imports-contacts-app sample
   - limiter for search methods
+  - search-results-paging-app sample
 
 ### Updated